+++ Enums/CommonGroup.cs
@@ -1,0 +1,8 @@
+﻿namespace Zeanon.Plugin.ArgusMonitor.Enums;
+
+public enum CommonGroup
+{
+    CPU_Temperature,
+    CPU_Multiplier,
+    CPU_Core_Clock,
+}