+++ Enums/CommonID.cs
@@ -1,0 +1,8 @@
+﻿namespace Zeanon.Plugin.ArgusMonitor.Enums;
+
+public enum CommonID
+{
+    CPU_Temperature_CPU_MAX,
+    CPU_Multiplier_CPU_MAX,
+    CPU_Core_Clock_CPU_MAX,
+}